--- a/schema.sql
+++ b/schema.sql
@@ -73,3 +73,7 @@ CREATE TABLE visits (
 ALTER TABLE owners ADD COLUMN email VARCHAR(120);
 
 CREATE INDEX index_visit_animal_id ON visits(animal_id);
+
+CREATE INDEX index_visit_vet_id ON visits(vet_id);
+
+CREATE INDEX index_owners_email ON owners(email);