### Database performance audit

In this Milestone, We:
- Add extra column to the owners table (email).
-  Add data to your database 
- Use EXPLAIN ANALYZE on queries to check what is happening
- Decrease the execution time of the first query
- Improve execution time of the other two queries.

🌊 SELECT COUNT(*) FROM visits where animal_id = 4;
     before adding index  :execution time - 1563ms
     after adding index:  execution time - 96.691ms
   
🌊SELECT * FROM visits where vet_id = 2;
    - before adding index :execution time - 1665ms
    - after adding index:  execution time - 992ms
    
🌊SELECT * FROM owners where email = '[owner_18327@mail.com](mailto:owner_18327@mail.com)';
     - before adding index :execution time - 590ms
     - after adding index:  execution time - [0.709ms](http://0.ms/) 

Find attached screenshots with EXPLAIN ANALYZE results;
![Explain](https://github.com/yashodhicy/VetClinic/assets/79149176/59b865c2-d2f5-42c8-af55-374716f79915)
![1st](https://github.com/yashodhicy/VetClinic/assets/79149176/123af330-8f03-4c09-b690-73f819e5f3f8)
![2nd](https://github.com/yashodhicy/VetClinic/assets/79149176/225019c2-d64f-4839-9eae-6685b915fcbb)
![3rd](https://github.com/yashodhicy/VetClinic/assets/79149176/461bca7e-914e-4b8f-b518-0c7f06eba0f9)