@Olamarx 2023-07-18
> Hi @Enning94,
> 
> Good job so far, indeed you have spent time on this project! There are some issues that you still need to work on to go to the next project but you are almost there! ![](https://camo.githubusercontent.com/957cdd5c6c73009f305aa460cba68fe9b38c4cf1eead69c6006a391516168166/68747470733a2f2f6d65646961312e67697068792e636f6d2f6d656469612f74385950663341434b656f754c7a454976722f67697068792e676966)
> 
> ## Highlights
> * Queries made ✔️
> * Readme file professional ✔️
> 
> ## Required Changes ♻️
> _Check the comments under the review._
> 
> * [ ]  I can see the first execution time is significantly more than 1 second, and the second query's execution time is more than the first and that is the opposite of what we want to achieve. The third, fourth and sixth are better but the fifth one is also against our aim. We are to make it faster than the previous. I will encourage you to play around with it again by using the index as there is no Index scan in the output and I can see Seq scan and Filter at the moment. Try that, let us see what our chances are in making it better.
> 
> ## Optional suggestions
> _Every comment with the [OPTIONAL] prefix is not crucial enough to stop the approval of this PR. However, I strongly recommend you to take them into account as they can make your code better._
> 
> Cheers and Happy coding!👏👏👏
> 
> Feel free to leave any questions or comments in the PR thread if something is not 100% clear. Please, remember to tag me in your question so I can receive the notification.
> 
> **_Please, do not open a new Pull Request for re-reviews. You should use the same Pull Request submitted for the first review, either valid or invalid unless it is requested otherwise._**
> 
> _As described in the [Code reviews limits policy](https://microverse.zendesk.com/hc/en-us/articles/1500004088561) you have a limited number of reviews per project (check the exact number in your Dashboard). If you think that the code review was not fair, you can request a second opinion using [this form](https://airtable.com/shrQAqnBwek5a0O0s)._

@Enning94 and @yashodhicy thanks for pulling up the call for clarity, I think the arrangement of the screenshots is the confusing factor here and I can see that the first three are the before while the after are the last three, I think you have done a good job. The next reviewer should be able to approve it.