@yashodhicy 2023-07-18
> @Enning94 and @yashodhicy thanks for pulling up the call for clarity, I think the arrangement of the screenshots is the confusing factor here and I can see that the first three are the before while the after are the last three, I think you have done a good job. The next reviewer should be able to approve it.
Thank you so much